fix database settings issues

+++ data/src/commonMain/kotlin/org/rhasspy/mobile/data/porcupine/PorcupineCustomKeyword.kt
@@ -8,6 +8,6 @@ import kotlinx.serialization.Serializable
 data class PorcupineCustomKeyword(
     val fileName: String,
     override val isEnabled: Boolean,
-    override val sensitivity: Float
+    override val sensitivity: Double
 ) : PorcupineKeyword
