+++ platformspecific/src/androidMain/kotlin/org/rhasspy/mobile/platformspecific/porcupine/PorcupineClient.kt
@@ -89,11 +89,11 @@ class PorcupineClient(
     private fun getSensitivities(): FloatArray {
         return wakeWordPorcupineKeywordDefaultOptions.filter { it.isEnabled && it.option.language == wakeWordPorcupineLanguage }
             .map {
-                it.sensitivity
+                it.sensitivity.toFloat()
             }.toMutableList().also { list ->
                 list.addAll(
                     wakeWordPorcupineKeywordCustomOptions.filter { it.isEnabled }.map {
-                        it.sensitivity
+                        it.sensitivity.toFloat()
                     }
                 )
             }.toTypedArray().toFloatArray()
