+++ settings/src/androidMain/kotlin/org/rhasspy/mobile/settings/settingsmigration/types/NewPorcupineKeywordCustomListSetting.kt
@@ -21,7 +21,7 @@ class NewPorcupineKeywordCustomListSetting(
                 PorcupineCustomKeyword(
                     fileName = it.value_,
                     isEnabled = it.enabled == 1L,
-                    sensitivity = it.sensitivity.toFloat()
+                    sensitivity = it.sensitivity
                 )
             }.toTypedArray().toImmutableList()
     }
@@ -34,7 +34,7 @@ class NewPorcupineKeywordCustomListSetting(
                         id = key.name,
                         value = it.fileName,
                         enabled = if (it.isEnabled) 1 else 0,
-                        sensitivity = it.sensitivity.toLong()
+                        sensitivity = it.sensitivity
                     )
             }
         }
